# api/docs/swagger.yaml
basePath: /
definitions:
  main.EvalResult:
    properties:
      flagKey:
        type: string
      isEnabled:
        type: boolean
      reason:
        type: string
    type: object
  main.Flag:
    properties:
      createdAt:
        type: string
      description:
        type: string
      isEnabled:
        type: boolean
      key:
        type: string
      rules:
        items:
          $ref: '#/definitions/main.Rule'
        type: array
    type: object
  main.FlagPatch:
    properties:
      description:
        type: string
      isEnabled:
        type: boolean
      rules:
        items:
          $ref: '#/definitions/main.Rule'
        type: array
    type: object
  main.Rule:
    properties:
      attribute:
        type: string
      operator:
        type: string
      rollout:
        type: integer
      type:
        type: string
      userIds:
        items:
          type: string
        type: array
      value:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: A lightweight feature flag service with rule-based evaluation.
  title: Feature Flag API
  version: "1.0"
paths:
  /flags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Flag'
            type: array
      summary: List all flags
      tags:
      - flags
    post:
      consumes:
      - application/json
      parameters:
      - description: Flag to create
        in: body
        name: flag
        required: true
        schema:
          $ref: '#/definitions/main.Flag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Flag'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a flag
      tags:
      - flags
  /flags/{key}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Flag key
        in: path
        name: key
        required: true
        type: string
      - description: Fields to update
        in: body
        name: flag
        required: true
        schema:
          $ref: '#/definitions/main.FlagPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Flag'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a flag
      tags:
      - flags
  /flags/{key}/evaluate:
    get:
      parameters:
      - description: Flag key
        in: path
        name: key
        required: true
        type: string
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      - description: User plan (e.g. enterprise)
        in: query
        name: plan
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.EvalResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Evaluate a flag for a user
      tags:
      - flags
  /health:
    get:
      description: Returns 200 if the service is up and the database is reachable.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Health check
      tags:
      - ops
  /metrics:
    get:
      description: Flag count, evaluation count, and last error seen — enough to spot
        trouble without reading logs.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Basic service metrics
      tags:
      - ops
swagger: "2.0"